+++ bindings/python/orderflow/_ffi.py
@@ -842,6 +842,14 @@ class OrderflowLib:
         ]
         lib.of_get_signal_descriptors_json.restype = c_int32
 
+        lib.of_get_signal_explanation_json.argtypes = [
+            c_void_p,
+            ctypes.POINTER(OfSymbol),
+            ctypes.POINTER(c_char_p),
+            ctypes.POINTER(c_uint32),
+        ]
+        lib.of_get_signal_explanation_json.restype = c_int32
+
         lib.of_compute_lob_features.argtypes = [
             c_void_p,
             ctypes.POINTER(OfSymbol),